refactor: usando numeric(10,2) em vez de Double no SQL

+++ sysagua-api/src/main/resources/db/migration/h2/V4__create_table_products.sql
@@ -1,8 +1,8 @@
 CREATE TABLE products (
     product_id BIGINT NOT NULL AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
-    price DOUBLE NOT NULL,
-    cost DOUBLE NOT NULL,
+    price NUMERIC(10,2) NOT NULL,
+    cost NUMERIC(10,2) NOT NULL,
     unit VARCHAR(255) NOT NULL,
     brand VARCHAR(255),
     category_id BIGINT,
